+++ src/utils/vscode/getInitialHTML.ts
@@ -79,6 +79,9 @@ export default function getInitialHTML(
  
         <h2>Need help?</h2>
         <p>Send an issue on <a href="https://github.com/watermelontools/wm-extension/issues">GitHub</a> and join us on <a href="https://join.slack.com/t/watermelonusers/shared_invite/zt-15bjnr3rm-uoz8QMb1HMVB4Qywvq94~Q">Slack</a></p>
+        <h2>Donate</h2>
+        <p><a href="https://github.com/sponsors/watermelontools">Help our Open Source efforts</a> by donating any amount.</p>
+ 
      </body>
      <footer>
       <script nonce="${nonce}" src="${scriptUri}" type="module"></script>