+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Added icons to the hover to make it more clear
 - Will now execute initial actions when you log in
+- Added a section to sponsor us from the sidebar
 
 ## [1.3.7]