+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added icons to the hover to make it more clear
 - Will now execute initial actions when you log in
 - Added a section to sponsor us from the sidebar
+- Will now show comment author profile pic for each comment
 
 ## [1.3.7]